+++ crates/bcr-ebill-api/src/service/contact_service.rs
@@ -168,7 +168,7 @@ impl ContactService {
         upload_id: &Option<String>,
         id: &NodeId,
         public_key: &PublicKey,
-        relay_url: &str,
+        relay_url: &url::Url,
         upload_file_type: UploadFileType,
     ) -> Result<Option<File>> {
         if let Some(upload_id) = upload_id {
@@ -198,7 +198,7 @@ impl ContactService {
         file_bytes: &[u8],
         node_id: &NodeId,
         public_key: &PublicKey,
-        relay_url: &str,
+        relay_url: &url::Url,
     ) -> Result<File> {
         let file_hash = util::sha256_hash(file_bytes);
         let encrypted = util::crypto::encrypt_ecies(file_bytes, public_key)?;