+++ crates/bcr-ebill-persistence/src/db/identity.rs
@@ -213,7 +213,7 @@ pub struct IdentityDb {
     pub country_of_birth: Option<Country>,
     pub city_of_birth: Option<String>,
     pub identification_number: Option<String>,
-    pub nostr_relays: Vec<String>,
+    pub nostr_relays: Vec<url::Url>,
     pub profile_picture_file: Option<FileDb>,
     pub identity_document_file: Option<FileDb>,
 }